# Conflicts:
#	protected/controllers/GpController.php

+++ protected/modules/OECaseSearch/controllers/CaseSearchController.php
@@ -4,12 +4,14 @@
  * Class CaseSearchController
  *
  * @property null|Trial $trialContext
+ * @property OECaseSearchModule $module
  */
+
 class CaseSearchController extends BaseModuleController
 {
     public $trialContext;
-    protected $parameters = array();
-    protected $parameterList = array();
+    protected array $parameters = array();
+    protected array $parameterList = array();
 
     public function filters()
     {
@@ -45,13 +47,15 @@ class CaseSearchController extends BaseModuleController
                     'renderPopups'
                 ),
                 'users' => array('@'),
+                'roles' => array('Advanced Search'),
             ),
         );
     }
 
     /**
      * Primary case search action.
-     * @param $trial_id int The Trial that this case search is in context of
+     * @param int|null $trial_id int The Trial that this case search is in context of
+     * @throws CException
      * @throws Exception
      */
     public function actionIndex($trial_id = null)
@@ -78,11 +82,27 @@ class CaseSearchController extends BaseModuleController
         $criteria = new CDbCriteria();
         $valid = $this->populateParams(true);
 
+        // Only run this next statement if not an advanced search super user.
+        if (!$this->checkAccess('TaskCaseSearchSuperUser')) {
+            $this->parameters[] = $this->buildParameterInstance(
+                999,
+                InstitutionParameter::class,
+                '=',
+                '{institution}',
+                false
+            );
+        }
+
         if ($valid && !empty($this->parameters)) {
             $this->actionClear();
-            $ids = array_column(Yii::app()->searchProvider->search($this->parameters), 'id');
+            /**
+             * @var $searchProvider SearchProvider
+             */
+            $searchProvider = Yii::app()->searchProvider;
+            $ids = array_column($searchProvider->search($this->parameters), 'id');
 
-            // Only copy to the $_SESSION array if it isn't already there - Shallow copy is done at the start if it is already set.
+            // Only copy to the $_SESSION array if it isn't already there.
+            // Shallow copy is done at the start if it is already set.
             if (!isset($_SESSION['last_search']) || empty($_SESSION['last_search'])) {
                 $_SESSION['last_search'] = $ids;
             }
@@ -94,7 +114,9 @@ class CaseSearchController extends BaseModuleController
         $criteria->with = 'contact';
         $criteria->compare('t.deleted', 0);
 
-        // A data provider is used here to allow faster search times. Results are iterated through using the data provider's pagination functionality and the CListView widget's pager.
+        // A data provider is used here to allow faster search times.
+        // Results are iterated through using the data provider's pagination functionality and the
+        // CListView widget's pager.
         $patientData = new CActiveDataProvider('Patient', array(
             'criteria' => $criteria,
             'totalItemCount' => count($ids),
@@ -128,7 +150,10 @@ class CaseSearchController extends BaseModuleController
             ),
         ));
 
-        $all_searches = SavedSearch::model()->findAll();
+        $all_searches = SavedSearch::model()->findAllByUserOrInstitution(
+            Yii::app()->user->id,
+            $this->selectedInstitutionId
+        );
 
         if (!empty($ids)) {
             foreach (Yii::app()->params['CaseSearch']['variables']['OECaseSearch'] as $var) {
@@ -177,14 +202,15 @@ class CaseSearchController extends BaseModuleController
 
         $this->render('index', array(
             'paramList' => $paramList,
-            'params' => (empty($this->parameters) && isset($_SESSION['last_search_params'])) ? $_SESSION['last_search_params'] : $this->parameters,
+            'params' => (empty($this->parameters) && isset($_SESSION['last_search_params']))
+                ? $_SESSION['last_search_params'] : $this->parameters,
             'patients' => $patientData,
             'patientsID' => $ids,
             'variables' => $variables,
             'variableList' => $variableList,
             'variableData' => $variable_data,
             'saved_searches' => $all_searches,
-            'search_label' => isset($_POST['search_name']) ? $_POST['search_name'] : '',
+            'search_label' => $_POST['search_name'] ?? '',
             'from_date' => $from ? $from->format('Y-m-d') : null,
             'to_date' => $to ? $to->format('Y-m-d') : null,
             'show_all_dates' => $show_all_dates,
@@ -196,7 +222,7 @@ class CaseSearchController extends BaseModuleController
      * @param $ids int[]|null
      * @return CaseSearchVariable
      */
-    protected function getVariableInstance($var, $ids = null)
+    protected function getVariableInstance($var, $ids = null): CaseSearchVariable
     {
         $variable = null;
 
@@ -221,22 +247,12 @@ class CaseSearchController extends BaseModuleController
     {
         $param = $_GET['parameter'];
 
-        /**
-         * @var $parameter CaseSearchParameter
-         */
-        $parameter = new $param['type']();
-        $parameter->id = $param['id'];
-        $parameter->operation = $param['operation'];
-        if (array_key_exists('value', $param)) {
-            if (is_array($param['value'])) {
-                foreach ($param['value'] as $value) {
-                    $key = $value['field'];
-                    $parameter->$key = $value['id'];
-                }
-            } else {
-                $parameter->value = $param['value'];
-            }
-        }
+        $parameter = $this->buildParameterInstance(
+            $param['id'],
+            $param['type'],
+            $param['operation'],
+            $param['value'] ?? null
+        );
 
         if ($parameter->validate()) {
             $this->renderPartial('parameter_form', array(
@@ -251,6 +267,42 @@ class CaseSearchController extends BaseModuleController
         }
 
         Yii::app()->end();
+    }
+
+    /**
+     * @param $id string
+     * @param $type string
+     * @param $operation string
+     * @param $value array|int|float|string
+     * @param $isSaved bool
+     * @return CaseSearchParameter
+     */
+    protected function buildParameterInstance(
+        string $id,
+        string $type,
+        string $operation,
+        $value,
+        $isSaved = true
+    ): CaseSearchParameter
+    {
+        /**
+         * @var $parameter CaseSearchParameter
+         */
+        $parameter = new $type();
+        $parameter->id = $id;
+        $parameter->operation = $operation;
+        if ($value) {
+            if (is_array($value)) {
+                foreach ($value as $valueEntry) {
+                    $key = $valueEntry['field'];
+                    $parameter->$key = $valueEntry['id'] ?? null;
+                }
+            } else {
+                $parameter->value = $value;
+            }
+        }
+        $parameter->isSaved = $isSaved;
+        return $parameter;
     }
 
     public function actionGetOptions()
@@ -271,14 +323,14 @@ class CaseSearchController extends BaseModuleController
      */
     public function actionLoadSearch($id)
     {
-        $preview = isset($_GET['preview']) ? $_GET['preview'] : null;
+        $preview = $_GET['preview'] ?? null;
 
         $search = SavedSearch::model()->findByPk($id);
         if (!$search) {
             throw new CHttpException(404, 'Saved search not found');
         }
         $this->actionClear();
-        $params = unserialize($search->search_criteria, array('allowed_classes' => true));
+        $params = unserialize($search->search_criteria);
         echo '<tbody>';
 
         foreach ($params as $param) {
@@ -327,11 +379,15 @@ class CaseSearchController extends BaseModuleController
                 /**
                  * @var $mergedParam CaseSearchParameter
                  */
-                $criteria_list[] = $parameter->saveSearch();
+                if ($parameter->isSaved) {
+                    // Only save parameters that are permitted to be saved.
+                    $criteria_list[] = $parameter->saveSearch();
+                }
             }
             $search_criteria = serialize($criteria_list);
             $search->search_criteria = $search_criteria;
             $search->name = $_POST['search_name'];
+            $search->institution_id = $this->selectedInstitutionId;
 
             if (!$search->save()) {
                 Yii::log(var_export($search->getErrors(), true));
@@ -346,13 +402,18 @@ class CaseSearchController extends BaseModuleController
      * Deletes the selected saved search
      * @param $id int ID of the saved search to delete.
      * @throws CHttpException
+     * @throws CDbException
      */
-    public function actionDeleteSearch($id)
+    public function actionDeleteSearch(int $id)
     {
         $search = SavedSearch::model()->findByPk($id);
 
         if (!$search) {
             throw new CHttpException(404, 'Unable to delete saved search - Saved search not found.');
+        }
+        // We want to ensure that only the creating user (or perhaps a super user) has permissions to delete a saved search.
+        if ($search->created_user_id !== Yii::app()->user->id) {
+            throw new CHttpException(403, 'You do not have permissions to delete this saved search.');
         }
         if (!$search->delete()) {
             throw new CHttpException(500, 'Unable to delete saved search - Unknown error occurred.');
@@ -361,7 +422,7 @@ class CaseSearchController extends BaseModuleController
 
     /**
      * Populate the list of search parameters.
-     * @param bool $populate_param_cache True if the controller's parameter cache should be populated as well; otherwise false.
+     * @param bool $populate_param_cache True if the parameter cache should be populated as well; otherwise false.
      * @return bool
      */
     protected function populateParams($populate_param_cache = false)
@@ -384,7 +445,9 @@ class CaseSearchController extends BaseModuleController
                          * @var $newParam CaseSearchParameter
                          */
                         $newParam = new $paramName();
-                        $newParam->attributes = $_POST[$paramName][$id];
+                        $attributes = $_POST[$paramName][$id];
+                        unset($attributes['type']);
+                        $newParam->attributes = $attributes;
                         if (!$newParam->validate()) {
                             $valid = false;
                         }
@@ -401,9 +464,6 @@ class CaseSearchController extends BaseModuleController
         $term = Yii::app()->request->getQuery('term');
         $type = Yii::app()->request->getQuery('type');
 
-        /**
-         * @var $type CaseSearchParameter
-         */
         $values = $type::getCommonItemsForTerm($term);
 
         $this->renderJSON($values);
@@ -411,10 +471,11 @@ class CaseSearchController extends BaseModuleController
 
     /**
      * Get the drilldown list for the selected datapoint.
-     * @param $patient_ids string List of patient IDs as a string (List is a string due to it being a parameter of a HTTP request).
+     * @param $patient_ids string List of patient IDs as a string (String because it is a parameter of a HTTP request).
      * @throws CException
      */
-    public function actionGetDrilldownList($patient_ids) {
+    public function actionGetDrilldownList(string $patient_ids)
+    {
         $pagination = array(
             'pageSize' => 10,
         );
@@ -492,7 +553,10 @@ class CaseSearchController extends BaseModuleController
         $this->populateParams();
         $ids = array_column(Yii::app()->searchProvider->search($this->parameters), 'id');
 
-        $variable = $this->getVariableInstance(Yii::app()->params['CaseSearch']['variables']['OECaseSearch'][$var], $ids);
+        $variable = $this->getVariableInstance(
+            Yii::app()->params['CaseSearch']['variables']['OECaseSearch'][$var],
+            $ids
+        );
 
         if (!isset($_POST['show-all-dates']) || $_POST['show-all-dates'] !== '1') {
             if (isset($_POST['from_date']) && $_POST['from_date']) {
@@ -513,7 +577,10 @@ class CaseSearchController extends BaseModuleController
 
     public function beforeAction($action)
     {
-        $assetPath = Yii::app()->assetManager->publish(Yii::getPathOfAlias('application.modules.OECaseSearch.assets'), true);
+        $assetPath = Yii::app()->assetManager->publish(
+            Yii::getPathOfAlias('application.modules.OECaseSearch.assets'),
+            true
+        );
         Yii::app()->clientScript->registerCssFile($assetPath . '/css/module.css');
 
         // This is required when the search results return any records.
@@ -522,12 +589,24 @@ class CaseSearchController extends BaseModuleController
             Yii::app()->assetManager->registerScriptFile('js/OpenEyes.UI.Dialog.js');
         }
 
-        if (!Yii::app()->clientScript->isScriptFileRegistered($assetPath . '/js/OpenEyes.UI.Dialog.LoadSavedSearch.js')) {
-            Yii::app()->assetManager->registerScriptFile('js/OpenEyes.UI.Dialog.LoadSavedSearch.js', 'application.modules.OECaseSearch.assets', -10);
+        if (!Yii::app()->clientScript->isScriptFileRegistered(
+            $assetPath . '/js/OpenEyes.UI.Dialog.LoadSavedSearch.js'
+        )) {
+            Yii::app()->assetManager->registerScriptFile(
+                'js/OpenEyes.UI.Dialog.LoadSavedSearch.js',
+                'application.modules.OECaseSearch.assets',
+                -10
+            );
         }
 
-        if (!Yii::app()->clientScript->isScriptFileRegistered($assetPath . '/js/OpenEyes.UI.Dialog.SaveSearch.js')) {
-            Yii::app()->assetManager->registerScriptFile('js/OpenEyes.UI.Dialog.SaveSearch.js', 'application.modules.OECaseSearch.assets', -10);
+        if (!Yii::app()->clientScript->isScriptFileRegistered(
+            $assetPath . '/js/OpenEyes.UI.Dialog.SaveSearch.js'
+        )) {
+            Yii::app()->assetManager->registerScriptFile(
+                'js/OpenEyes.UI.Dialog.SaveSearch.js',
+                'application.modules.OECaseSearch.assets',
+                -10
+            );
         }
 
         return parent::beforeAction($action);
